fix CRUD operation

+++ src/main/java/ru/esplit/first_security_app/services/AdminServiceImpl.java
@@ -46,8 +46,11 @@ public class AdminServiceImpl implements AdminService {
 
     @Transactional
     @Override
-    public void delete(long id) {
-        peopleRepository.deleteById(id);
+    public void delete(long id, long principalId) {
+        if (id != principalId) {
+            peopleRepository.deleteById(id);
+        }
+        
     }
 
     @Transactional
